Update UI to warn user of data duplication potential
If a user uploads a csv file with some bad rows, the good rows
will be processed but not the bad rows.

Give the user some prompting so they don't re-upload the entire
file, which would cause the good rows to be duplicated.

Ellen will create a new Issue to reject the entire CSV when there
are bad rows, so the user can reupload the entire CSV as one piece.