simplify CSV importing
Because the db/seeds and the ImportJob jobs were both importing CSVs, it
made sense to consolidate the business logic of importing CSV data in
one place. The CsvImporter seems a good place.

* Made CsvImporter responsible for doing the actual import

This makes the background jobs a lot simpler by deferring to the importer
to do the actual data processing.

CsvImporter refactored from a module to a class so that an individual
import action is represented by an instance of an importer. A class
method import() remains that maintains the previous method signature so
the existing call sites for CSV import stay the same.

* CsvImporter::CATEGORIES

This is a simpler list of the human-readable strings that are the
categories of data to import. This is the One True Place™ where
categories that are importable by CSV are declared. Everywhere that does
imports (the import jobs and the seeds) refer to this list to know
what's importable.

* Import statistics generation moved to CsvImporter

* CSV-to-database-columns translation moved to the models

The create_from_csv_data() on the models is now responsible for
using a row of CSV data and instantiating a new model ready to be saved
to the database. A future improvment to this method could be to perform
any data cleansing currently handled by data_cleanse!() and save the
model to the database.

The HEADERS mapping between CSV column names and database column names
already lived in the models.  Colocating the mapping and the methods
that execute on that mapping should make maintaining them clearer.

Signed-off-by: Robb Kidd <robb@thekidds.org>