Skip failing test that tests the wrong thing
This test fails on CI because the csv file to be processed does not exist.
Can repro locally by first running  a `rake tmp:cache:clear`

Skipping because at the same time, Ellen noticed that workers on heroku
also cannot process csv files that are local to the web worker disk.
So this test is testing a process we cannot use in production anyway.