move uploaded file deletion to only after import attempted
Statistics don't exist until a CSV import is attempted, so move the file
deletion to after the records are imported when the headers are valid.

Because the file is being kept when some rows were not imported, I have
assumed that the file is kept when there are errors. Therefore,
remote_file! is not also called for the invalid headers branch.

Signed-off-by: Robb Kidd <robb@thekidds.org>