Create Animals, Consolidation Reports, and Familes. (#130)
* Create Animals, Consolidation Reports, and Familes.

* Rename migration files to solve family relation error

* Fix family relationships.

The relationships between models didn't work. Mobbed with Betsy Jennifer Linda Bart.

We moved the belongs_to from Animal to Family because we want Familes to know their Male/Female Animals, but we don't need Animals to know their Families.

We didn't set up the corresponding has_many from Animals to Families, because it looked like that would be complicated to set up, and it's not a necessary relationship for us to define right now.

Note to future devs: put off setting up has_many :animals for as long as possible, because it will be complex.

* Add relational specs.

Got this awesome pattern from Linda from CASA.

Also thank you to Bart!

* Add spec tests for:

Consolidation Report, Facility, Family, and Tank Associations.