store process pid files on a temporary filesystem
Mounting the tmp/pids directory to a tmpfs volume will ensure that when
the container is stopped, the pid files go away, too. This should take
care of the pid file cleanup, so removing the entrypoint script that
existed solely to do pid file cleanup.

The web service in the compose file can now rely on the Rails service
startup CMD in the Dockerfile, so it does not need an overridden
"command:".