🚧  WIP: A Very Docker Development Environment (#226)
* store db data on an "internal" docker volume

Docker will establish a shared volume named "dbdata" that persists
between instantiations of the "db" container. Development data will
stick around until volume(s) are removed with:

    docker-compose down --volumes

Mounting the data directory for the database this way has a couple
benefits.

+ It's pretty easy to clear out the development and test databases with
  a single docker-compose command. Good for fresh starts.
+ For macOS and Windows, database disk operations are faster. Disk
  access for database files remain on the virtual machine that is what
  actually runs Docker on Windows and masOS. Now that disk I/O is not
  traversing the slower filesystem mount to the host OS.

Added a Makefile and a "nuke" target for development environments that
have "make" available. Now all running containers can be stopped,
removed, and any disposable volumes deleted with one command: make nuke.

* make db service responsible for its readiness to the app

The 2.x schema, while seemingly old, has some useful features for
development environments. Particularly the "condition:" key within a
"depends_on:" statement within a service.

The 2.4 docker-compose format is still supported for the foreseeable
future, even if it is not receiving new features. 2.x is recommended
when using docker-compose only for development environments.

The official postgres image supports custom initialization behavior[1]
when it sees that there is no data in its data directory (i.e. first
time running, maybe after its data volume was deleted for a fresh
start). Place scripts in the container's /docker-entrypoint-initdb.d
directory and they will be run in alphanumerical order by the postgres
entrypoint after a new database has been initialized.

These changes do the following:

+ changes the compose file schema version to 2.x
+ a new SQL script added to ./bin/dbinit that creates _development
   and _test databases
+ the ./bin/dbinit directory mounted into the "db" container at the
   location used by the entrypoint script.
+ added a healthcheck test to the "db" service to define "healthy" as
   the database service is serving up a database named
   "abalone_development" on the default TCP port

This means that "docker-compose up db" ought to result in the
abalone_development (and _test) databases being ready for schema
migration and running the application.

With "healthiness" defined for the database, the web and delayed_job
containers can now be told to wait to start until "db" is healthy. No
need for a wait_for_it.sh script here in development.

[1] See "Initialization scripts" section of the official postgres image
    README. https://hub.docker.com/_/postgres

* use YAML merge type to reduce duplication

Add a new service definition "app" to roll up the things that are common
between running the Rails web process and the Rails delayed_job process.

+ the Railsy services build from the same Dockerfile.
+ the Railsy servicers need the Rails root mounted up to see local dev
   changes.
+ the Railsy services need the database to be available

Only the differences are left:

+ web is the only service that needs ports forwarded
+ delayed_job has a different start command

* only listen on localhost in development

"3000:3000" binds the web app container's port 3000 to the host's port
3000 on *all* of the host's network interfaces. Meaning you might be
serving up your development app in all its maybe-insecure glory to the
coffee shop wifi.

While not a big problem currently, when we start hanging out in coffee
shops again, you'll want to be sharing this web service only to yourself.

* give the image a name in docker-compose

Bonus "make build" target to use docker-compose to build the abalone
image.

* mount the app tmp directory to a shared docker volume

This solves issues where tmp/cache has things on the host disk that the
Rails process in the container can't modify or is upset about already
being present.

* store process pid files on a temporary filesystem

Mounting the tmp/pids directory to a tmpfs volume will ensure that when
the container is stopped, the pid files go away, too. This should take
care of the pid file cleanup, so removing the entrypoint script that
existed solely to do pid file cleanup.

The web service in the compose file can now rely on the Rails service
startup CMD in the Dockerfile, so it does not need an overridden
"command:".

* set the Rails services to run in development mode

* super duper convenience!

make targets and compose service definitions for routine tasks

Signed-off-by: Robb Kidd <robb@thekidds.org>