default the db service to listen on host's localhost:5432 (#228)
Forward the "db" service's port to the host's loopback interface, so
that a developer could choose to use docker-compose only for a container
to run the database while running all the Ruby processes on their host
computer. "127.0.0.1:5432:5432" was chosen over "5432:5432" so that the
PostgreSQL would not be available to all other computers on the host
computer's network (say, a coffee shop wifi).

This also updates the development database config to default to using
"localhost" for the database host to support a typical dev environment
where Ruby processes are running on the host OS. It accepts an
environment variable "ABALONE_DATABASE_HOSTNAME" (named in the style of
the existing "ABALONE_DATABASE_PASSWORD" variable) to override the
database hostname. The compose file has been updated to override with
the hostname of the composed "db" service.

The test database config really only differs from the development
configuration

Risk: port conflicts on developer computers who
already have Postgres for Windows or Postgres.app installed and might
not know how or want to stop or uninstall those services.

Signed-off-by: Robb Kidd <robb@thekidds.org>