Removed the dev env db username/password
Having this constraint forces developers to have the username postres
with a specified username. This can cause conflicts with other
applications that require a username postgres and a different password.
It also causes unessesary install issues for those who are less familiar
with pg. I suggest removing this constraint all together for the dev
environment.