fix db tasks being upset when DATABASE_URL is set
You wouldn't think you need to do this, but adding the url: key if there
is a URL set in the environment ensures that the URL is used to merge
the test DB config when rake db:* tasks try to manipulate both
development and test databases. A db name given in the test: stanza of
database.yml will still override the db name given in the URL. Trying to
manipulate both dev and test databases when there is a URL was turned
off in rails/rails#37351, but that hasn't been released yet.

Signed-off-by: Robb Kidd <robb@thekidds.org>