Rework measurements to have correct associations and values (#379)
Rework measurements to have correct associations and values

Add a measurement type model for tracking different measurements per
organization for different values (e.g. length, count) and units (e.g. cm,
integer)
Move around associations to have a polymorphic assocations for the test
subject, and measurements, measurement types and measurement events to
belong to each other correctly.
Remove tank_id from measurement event as they could belong to many
different things which you can diverge from the associated measurements.
Add specs for all of the above.
Deprecate specs for CSVs and File Uploads, and assocaited methods on
models, such as #create_from_csv . The reason for this is that they are
going to be entirely reworked, and the current usage of them only
facilities specs, and not provides functionality to the app. I'm
expecting all these skipped tests will be removed further down the line
when the CSV import is reworked.


Co-authored-by: Brian Todt <todtb@users.noreply.github.com>
Co-authored-by: Sebastian Kopp <smkopp92@gmail.com>